+++ Sharpify/AssemblyInfo.cs
@@ -1,4 +1,4 @@
 ﻿using System.Reflection;
 
-[assembly: AssemblyFileVersion("1.4.0")]
-[assembly: AssemblyVersion("1.4.0")]
+[assembly: AssemblyFileVersion("1.4.1")]
+[assembly: AssemblyVersion("1.4.1")]